Added Intro Scene, Intro Music, GUI Text, Started Script to Show Text
Having roughly timed how long it would take a person to read through
the paragraphs of text in the intro, I’ve created a song and added it
to the Music Assets folder.

I’ve also written out the script in several GUI text paragraphs, added
it to the Intro Scene, and started a C# script that will turn each
paragraph on/off after a certain number of seconds have elapsed.

However, I have not yet connected the music or the C# script (the
script is not yet working, so the non-working parts are commented out)