Updated timing of intro texts to match music
Rearranged a couple paragraphs in the intro so that they’d be easier to
time with the music. Re-named a couple tags to go along with the new
paragraphs, and updated them in the C# script.

Also, updated the number of seconds between paragraph transitions so
that they paragraph transitions would fit with the music transitions.