Swapped Levels 2 & 3, Move Coke Machines
I swapped levels 2 and 3. Move all coke machines to either z=6 or z=-1
I also changed all velocity multitudes to 15 so they can get the extra
reach they need to make it to the end of the maps. Cannot get the Coke
Cans to spawn in the middle of the coke machines. If I change the x
value of where they spawn, it shifts it automatically to the left or
right side regardless of what the x value is