Some trouble: Added Medal Music, Got Transitions to Fire After Success Text is Visible
Made lots of changes so that the transitions between levels are better.

However, while the transitions WERE working in individual scenes, when
playing the game from the Intro forward, the success text seems to show
all the time, and the golf cart is only capable of lateral movement on
some scenes.  Not sure what exactly went wrong here. Need to figure
that out.